get rid of old LevelFeedbackEntry and update new components, start updating tests
